# src/main/java/org/vb/model/RemoteApiStatus.java
/*
 * Country Code Lookup API
 * API for conversion country code to country name from remote API https://www.travel-advisory.info/api
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package org.vb.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import org.vb.model.RemoteApiStatusApiStatus;
import javax.validation.constraints.*;

/**
 * Remote Api Status JSON file format.
 */
@ApiModel(description = "Remote Api Status JSON file format.")
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2020-10-21T21:41:38.212-07:00")
public class RemoteApiStatus   {
  @JsonProperty("api_status")
  private RemoteApiStatusApiStatus apiStatus = null;

  public RemoteApiStatus apiStatus(RemoteApiStatusApiStatus apiStatus) {
    this.apiStatus = apiStatus;
    return this;
  }

  /**
   * Get apiStatus
   * @return apiStatus
   **/
  @JsonProperty("api_status")
  @ApiModelProperty(value = "")
  public RemoteApiStatusApiStatus getApiStatus() {
    return apiStatus;
  }

  public void setApiStatus(RemoteApiStatusApiStatus apiStatus) {
    this.apiStatus = apiStatus;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    RemoteApiStatus remoteApiStatus = (RemoteApiStatus) o;
    return Objects.equals(this.apiStatus, remoteApiStatus.apiStatus);
  }

  @Override
  public int hashCode() {
    return Objects.hash(apiStatus);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class RemoteApiStatus {\n");
    
    sb.append("    apiStatus: ").append(toIndentedString(apiStatus)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}